+++ apps/web/sections/NFTCollectionCarousel/assets/arrow-left-icon.tsx
@@ -1,0 +1,21 @@
+export const ArrowLeftIcon = ({ className }: { className?: string }) => {
+  return (
+    <svg
+      width="19"
+      height="35"
+      viewBox="0 0 19 35"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M18.117 33.6734C17.5312 34.2591 16.5815 34.2591 15.9957 33.6734L0.43934 18.117C-0.146447 17.5312 -0.146447 16.5815 0.43934 15.9957C1.02513 15.4099 1.97487 15.4099 2.56066 15.9957L18.117 31.552C18.7028 32.1378 18.7028 33.0876 18.117 33.6734Z"
+        fill="#D9D9D9"
+      />
+      <path
+        d="M0.43934 18.117C-0.146447 17.5312 -0.146447 16.5815 0.43934 15.9957L15.9957 0.43934C16.5815 -0.146446 17.5312 -0.146447 18.117 0.43934C18.7028 1.02513 18.7028 1.97487 18.117 2.56066L2.56066 18.117C1.97487 18.7028 1.02513 18.7028 0.43934 18.117Z"
+        fill="#D9D9D9"
+      />
+    </svg>
+  );
+};
